pyproject.toml
[project]
name = "rag-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "pymilvus>=2.5.4",
    "google-generativeai>=0.8.4",
]
